requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
openpyxl
google-generativeai
psycopg2-binary
SQLAlchemy
python-dotenv
Jinja2
